Re-delete duplicate components
