+++ frontend/src/components/transactions/TransactionForm.jsx
@@ -17,15 +17,59 @@ export default function TransactionForm({ onSubmit, onCancel }) {
 
   // Handle sudden change in transaction type
   useEffect(() => {
-    if(amount === ""){
-      return
+
+    // Disable submit button if title or amount is empty/invalid
+    if(title === "" || amount === "" || Math.abs(amount) === 0 ){
+      setDisableClick(true);
+      disableClickSync.current = true;
+      return;
+    } else {
+      setDisableClick(false);
+      disableClickSync.current = false;
     }
+
+    // Ensure amount is positive for incomes and negative for expenses
     if (transactionType === "expense") {
       setAmount(-Math.abs(amount));
     } else {
       setAmount(Math.abs(amount));
     }
-  }, [transactionType, amount]);
+  }, [transactionType, title, amount]);
+
+  const handleAmountChange = (e) => {
+    let value = e.target.value;
+
+    if (value === "") {
+      setAmount("");
+      return;
+    }
+
+    // Split into integer and decimal parts
+    const [integerPart, decimalPart] = value.split(".");
+    
+    // the below comment is required for the BigInt working
+    /* global BigInt */
+    try {
+      // Check if the integer part exceeds the allowed digits (14 digits)
+      const integerBigInt = BigInt(integerPart);
+      if (integerBigInt > BigInt("99999999999999") || integerBigInt < BigInt("-99999999999999")) {
+        alert("Amount must be between -99,999,999,999,999.99 and +99,999,999,999,999.99.");
+        return;
+      }
+
+      // Handle decimal part (ensure it is not more than 2 decimal places)
+      if (decimalPart && decimalPart.length > 2) {
+        alert("Amount cannot have more than 2 decimal places.");
+        return;
+      }
+
+      // If valid, set the amount
+      setAmount(value);
+    } catch (error) {
+      alert("Please enter a valid number.");
+    }
+  };
+
 
   // Prevent form submission on enter key press
   const handleKeyDown = (event) => {
@@ -113,13 +157,8 @@ export default function TransactionForm({ onSubmit, onCancel }) {
             <input
               type="number"
               id="amount"
-
-              
               value={amount}
-              
-              onChange={(e) => {setAmount(e.target.value)}
-              }
-
+              onChange={handleAmountChange}
               onKeyDown={handleKeyDown}
               className={inputStyle}
             ></input>
