+++ frontend/src/components/transactions/TransactionLog.jsx
@@ -4,9 +4,12 @@ import { IoIosArrowForward, IoIosArrowBack } from "react-icons/io";
 import TransactionContext from "../../context/TransactionContext";
 import { useContext, useState, useEffect } from "react";
 
+import { LoadingSpinner } from "../LoadingSpinner";
+
 export default function TransactionList() {
   const {
     transactions,
+    allTransactions,
     filter,
     currentPage,
     setCurrentPage,
@@ -17,28 +20,45 @@ export default function TransactionList() {
   const [maxPage, setMaxPage] = useState(100);
   const [isPageJustLoaded, setIsPageJustLoaded] = useState(true);
   const [isFiltering, setIsFiltering] = useState(false);
+  const [loading, setLoading] = useState(false);
+
+
 
   // at page load, transactions is empty, so set maxPage to currentPage (1). So using isPageJustLoaded to avoid this behavior at page load
   // only set the max page after the transactions have been loaded
   useEffect(() => {
+    setMaxPage(Math.ceil(allTransactions.length / 10));
     if (!isPageJustLoaded && !isFiltering) {
-      if (transactions.length == 0) {
+      if (transactions.length < 10) {
         setMaxPage(currentPage);
+        
       }
     } else {
       setIsPageJustLoaded(false);
       setIsFiltering(false);
     }
-  }, [transactions,]);
+
+  }, [transactions, allTransactions, currentPage, isPageJustLoaded, isFiltering]);
 
   useEffect(() => {
     if (filter.length > 0) {
       setIsFiltering(true);
     } else {
       setIsFiltering(false);
-      setMaxPage(100);
     }
   }, [filter]);
+
+  useEffect(() => {
+    const fetchTransactions = () => {
+      if (loading){
+        setLoading(false);
+      }
+  
+    };
+
+    fetchTransactions();
+  }, [transactions]);
+
 
   return (
     <div className="flex flex-col items-center">
@@ -87,7 +107,9 @@ export default function TransactionList() {
         </div>
         <div className=" flex justify-between flex-col items-center min-h-[450px] outline outline-4 outline-primary rounded-3xl mt-4 pb-3">
           <div className="w-[90%] mt-[30px]">
-            {transactions.length !== 0 ? (
+            {loading ? (
+              <LoadingSpinner /> // Show spinner while loading
+            ) : transactions.length !== 0 ? (
               <ul>
                 {transactions.map((transaction) => (
                   <Transaction
@@ -98,41 +120,39 @@ export default function TransactionList() {
                 ))}
               </ul>
             ) : (
-              <div className="flex flex-col items-center gap-6">
-                <h1 className="text-body">No More Transactions To Load...</h1>
-                <button
-                  className="text-button-small text-white bg-primary rounded-full px-5 py-1 hover:bg-primary-dark active:bg-primary-darker"
-                  onClick={() => setCurrentPage(1)}
-                >
-                  Go back
-                </button>
-              </div>
+              <LoadingSpinner />
             )}
           </div>
 
           <div className="flex flex-row min-w-[105px] justify-between">
             <button
-              className={
-                currentPage === 1
+              className={currentPage === 1
                   ? "text-gray-400"
                   : "text-black hover:text-primary active:text-primary-dark"
               }
-              onClick={() => setCurrentPage(currentPage - 1)}
+              onClick={() => {setCurrentPage(currentPage - 1)
+                setLoading(true);
+              }}
               disabled={currentPage === 1}
+
             >
-              <IoIosArrowBack size={35} />
+              {( <IoIosArrowBack size={35} /> // Normal content when not loading
+              )}
             </button>
             <h2 className="text-sub-heading">{currentPage}</h2>
+            <h3 className="text-sub-heading">&#47;{maxPage}</h3>
             <button
-              className={
-                currentPage === maxPage
+              className={currentPage === maxPage
                   ? "text-gray-400"
                   : "text-black hover:text-primary active:text-primary-dark"
               }
-              onClick={() => setCurrentPage(currentPage + 1)}
+              onClick={() => {setCurrentPage(currentPage + 1)
+                setLoading(true);
+              }}
               disabled={currentPage === maxPage}
             >
-              <IoIosArrowForward size={35} />
+              {( <IoIosArrowForward size={35} /> // Normal content when not loading
+              )}
             </button>
           </div>
         </div>
